Cache images + client jwt
Generate thumbnails on the file or pre-generate them with
the provided thumbnail command.

Redirect client to sign in page if error happened. It can
happen mostly because the JWT token expired. Anyway,
if it was not able to load the object list, it's better to
just redirect to the sign in page.

resolves #2
resolves #1